+++ extensions/foo_jl_queue_manager_mac/src/UI/QueueRowView.h
@@ -9,5 +9,8 @@
 
 #import <Cocoa/Cocoa.h>
 
-@interface QueueRowView : NSTableRowView
+@interface QueueRowView : NSTableRowView {
+    BOOL _cachedTransparentMode;
+    BOOL _transparentModeCached;
+}
 @end